Change datatype default to np.float32 everywhere instead of passing in np.float32 everywhere

+++ npbench/benchmarks/go_fast/go_fast.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 
-def initialize(N, datatype=np.float64):
+def initialize(N, datatype=np.float32):
     from numpy.random import default_rng
     rng = default_rng(42)
     x = rng.random((N, N), dtype=datatype)
